+++ lib/controllers/penyuluh/palawija_controller.dart
@@ -136,14 +136,16 @@ class PalawijaController {
         return;
       }
 
-      final db = await DatabaseHelper().database;
-      await db.delete(
-        'detailPalawija',
-        where: "id = ?",
-        whereArgs: [id],
-      );
+      if (result) {
+        final db = await DatabaseHelper().database;
+        await db.delete(
+          'detailPalawija',
+          where: "id = ?",
+          whereArgs: [id],
+        );
 
-      EasyLoading.showToast("Berhasil menghapus data");
+        EasyLoading.showToast("Berhasil menghapus data");
+      }
     } catch (e) {
       EasyLoading.showToast("Gagal menghapus data");
       log("Delete error: $e");
